+++ R/app_server.R
@@ -51,7 +51,8 @@ app_server <- function(input = NULL, output = NULL, session = NULL) {
     "filter_data" = get_config("filter_data"),
     "filter_key" = get_config("filter_key"),
     "startup_msg" = get_config("startup_msg"),
-    "reload_period" = get_config("reload_period")
+    "reload_period" = get_config("reload_period"),
+    "enable_dataset_filter" = get_config("enable_dataset_filter")
   )
 
   app_server_(input, output, session, opts)
@@ -85,6 +86,7 @@ app_server_ <- function(input, output, session, opts) {
   )
 
   module_server <- opts[["module_info"]][["server_list"]]
+  module_meta <- opts[["module_info"]][["meta_list"]]
   module_names <- opts[["module_info"]][["module_name_list"]]
   module_hierarchy_list <- opts[["module_info"]][["hierarchy_list"]]
   data <- opts[["data"]]
@@ -92,6 +94,7 @@ app_server_ <- function(input, output, session, opts) {
   filter_key <- opts[["filter_key"]]
   startup_msg <- opts[["startup_msg"]]
   reload_period <- opts[["reload_period"]]
+  enable_dataset_filter <- opts[["enable_dataset_filter"]]
 
   datasets_filters_info <- get_dataset_filters_info(data, filter_data)
 
@@ -131,62 +134,120 @@ app_server_ <- function(input, output, session, opts) {
     shiny::reactive(unfiltered_dataset()[[filter_data]])
   )
 
-  dataset_filters <- local({
-    l <- vector(mode = "list", length = length(datasets_filters_info))
-    names(l) <- names(datasets_filters_info)
-    for (idx in seq_along(datasets_filters_info)) {
-      l[[idx]] <- local({
-        curr_dataset_filter_info <- datasets_filters_info[[idx]]
-        dv.filter::data_filter_server(
-          curr_dataset_filter_info[["id"]],
-          shiny::reactive({
-            unfiltered_dataset()[[curr_dataset_filter_info[["name"]]]] %||% data.frame()
-          })
-        )
-      })
-    }
 
-    l
-  })
+  if (enable_dataset_filter) {
+    log_inform("Dataset filter server")
 
-  filtered_dataset <- shinymeta::metaReactive({
-    # dv.filter returns a logical vector. This contemplates the case of empty lists
-    shiny::req(is.logical(global_filtered_values()))
+    dataset_filters <- local({
+      l <- vector(mode = "list", length = length(datasets_filters_info))
+      names(l) <- names(datasets_filters_info)
+      for (idx in seq_along(datasets_filters_info)) {
+        l[[idx]] <- local({
+          curr_dataset_filter_info <- datasets_filters_info[[idx]]
+          dv.filter::data_filter_server(
+            curr_dataset_filter_info[["id"]],
+            shiny::reactive({
+              unfiltered_dataset()[[curr_dataset_filter_info[["name"]]]] %||% data.frame()
+            })
+          )
+        })
+      }
 
-    # Depend on all datasets
-    purrr::walk(dataset_filters, ~ .x())
+      l
+    })
 
-    # We do not react to changed in unfiltered dataset, otherwise when a dataset changes
-    # We filter the previous dataset which in the best case produces and extra reactive beat
-    # and in the worst case produces an error in (mvbc)
-    # We don't want to control the error in (mvbc) because filtered dataset only changes when filter changes
-    ufds <- shiny::isolate(unfiltered_dataset())
+    filtered_dataset <- shinymeta::metaReactive({
+      # dv.filter returns a logical vector. This contemplates the case of empty lists
+      shiny::req(is.logical(global_filtered_values()))
 
-    curr_dataset_filters <- dataset_filters[intersect(names(dataset_filters), names(ufds))]
+      # Depend on all datasets
+      purrr::walk(dataset_filters, ~ .x())
 
-    # Current dataset must be logical with length above 0
-    # Check dataset filters check all datafilters are initialized
-    purrr::walk(curr_dataset_filters, ~ shiny::req(checkmate::test_logical(.x(), min.len = 1)))
+      # We do not react to changed in unfiltered dataset, otherwise when a dataset changes
+      # We filter the previous dataset which in the best case produces and extra reactive beat
+      # and in the worst case produces an error in (mvbc)
+      # We don't want to control the error in (mvbc) because filtered dataset only changes when filter changes
+      ufds <- shiny::isolate(unfiltered_dataset())
 
-    filtered_key_values <- ufds[[filter_data]][[filter_key]][global_filtered_values()]
+      curr_dataset_filters <- dataset_filters[intersect(names(dataset_filters), names(ufds))]
 
-    fds <- ufds
+      # Current dataset must be logical with length above 0
+      # Check dataset filters check all datafilters are initialized
+      purrr::walk(curr_dataset_filters, ~ shiny::req(checkmate::test_logical(.x(), min.len = 0)))
 
-    # Single dataset filtering
-    fds[names(curr_dataset_filters)] <- purrr::imap(
-      fds[names(curr_dataset_filters)],
-      function(val, nm) {
-        # (mvbc)
-        fds[[nm]][dataset_filters[[nm]](), , drop = FALSE]
+      filtered_key_values <- ufds[[filter_data]][[filter_key]][global_filtered_values()]
+
+      fds <- ufds
+
+      # Single dataset filtering
+      fds[names(curr_dataset_filters)] <- purrr::imap(
+        fds[names(curr_dataset_filters)],
+        function(val, nm) {
+          # (mvbc)
+          labels <- get_lbls(fds[[nm]])
+          current_fds <- fds[[nm]][dataset_filters[[nm]](), , drop = FALSE]
+          set_lbls(current_fds, labels)
+        }
+      )
+
+      # Global dataset filtering
+      global_filtered <- purrr::map(
+        fds, function(current_ds) {
+          mask <- current_ds[[filter_key]] %in% filtered_key_values
+          labels <- get_lbls(current_ds)
+          current_ds <- current_ds[mask, , drop = FALSE]
+          set_lbls(current_ds, labels)
+        }
+      )
+    })
+
+    tab_ids <- c("__tabset_0__", names(opts[["module_info"]][["tab_group_names"]]))
+    shiny::observeEvent(
+      {
+        purrr::map(tab_ids, ~ input[[.x]])
+      },
+      {
+        current_tab <- "__tabset_0__"
+        zero_tabs <- length(input[["__tabset_0__"]]) == 0
+        if (!zero_tabs) {
+          while (!current_tab %in% opts[["module_info"]][["module_id_list"]]) {
+            current_tab <- input[[current_tab]]
+          }
+        }
+
+        used_ds <- used_datasets[[current_tab]]
+        all_nm <- names(datasets_filters_info)
+        if (!zero_tabs && !is.null(used_ds)) {
+          used_nm <- intersect(used_datasets[[current_tab]], names(datasets_filters_info))
+          unused_nm <- setdiff(all_nm, used_nm)
+        } else {
+          used_nm <- all_nm
+          unused_nm <- character(0)
+        }
+
+        for (nm in unused_nm) {
+          shinyjs::hide(datasets_filters_info[[nm]][["id_cont"]])
+        }
+
+        for (nm in used_nm) {
+          shinyjs::show(datasets_filters_info[[nm]][["id_cont"]])
+        }
       }
     )
+  } else {
+    log_inform("Single filter server")
 
-    # Global dataset filtering
-    global_filtered <- purrr::map(
-      fds,
-      ~ dplyr::filter(.x, .data[[filter_key]] %in% filtered_key_values) # nolint
-    )
-  })
+    filtered_dataset <- shinymeta::metaReactive({
+      # dv.filter returns a logical vector. This contemplates the case of empty lists
+      shiny::req(is.logical(global_filtered_values()))
+      log_inform("New filter applied")
+      filtered_key_values <- unfiltered_dataset()[[filter_data]][[filter_key]][global_filtered_values()] # nolint
+      purrr::map(
+        unfiltered_dataset(),
+        ~ dplyr::filter(.x, .data[[filter_key]] %in% filtered_key_values) # nolint
+      )
+    })
+  }
 
   # Prepare module_output argument
   module_output_env <- rlang::current_env()
@@ -257,10 +318,19 @@ app_server_ <- function(input, output, session, opts) {
     )
   )
 
+  used_datasets <- list()
+
   module_output <- list()
   for (srv in module_server) {
-    module_output[[srv[["module_id"]]]] <- srv[["server"]](module_args)
+    mod_id <- srv[["module_id"]]
+    srv_fun <- srv[["server"]]
+
+    module_output[[mod_id]] <- srv_fun(module_args)
+    used_datasets[[mod_id]] <- module_meta[[mod_id]][["meta"]][["dataset_info"]][["all"]]
   }
+
+
+
 
   #### Report modal