+++ tests/testthat/test-app_server.R
@@ -16,7 +16,8 @@ test_that(
       data = datasets,
       filter_data = "a",
       module_info = resolve_module_list(list()),
-      filter_key = "mpg"
+      filter_key = "mpg",
+      enable_dataset_filter = TRUE
     )
 
 
@@ -45,7 +46,8 @@ test_that(
       data = datasets,
       filter_data = "a",
       module_info = resolve_module_list(list()),
-      filter_key = "mpg"
+      filter_key = "mpg",
+      enable_dataset_filter = TRUE
     )
 
     shiny::testServer(app_server_test(testing_options), {
@@ -63,7 +65,8 @@ test_that(
   {
     testing_options <- list(
       data = list(),
-      module_info = resolve_module_list(list("mod_1" = mod_identity(1, "mod_1"), "mod_2" = mod_identity(2, "mod_2")))
+      module_info = resolve_module_list(list("mod_1" = mod_identity(1, "mod_1"), "mod_2" = mod_identity(2, "mod_2"))),      
+      enable_dataset_filter = TRUE
     )
 
     shiny::testServer(app_server_test(testing_options), {
@@ -84,7 +87,8 @@ test_that(
       data = datasets,
       filter_data = "a",
       module_info = resolve_module_list(list()),
-      filter_key = "mpg"
+      filter_key = "mpg",
+      enable_dataset_filter = TRUE
     )
 
     shiny::testServer(app_server_test(testing_options), {
@@ -104,7 +108,8 @@ test_that(
       data = list(),
       filter_data = NULL,
       module_info = resolve_module_list(list("mod_1" = mod_identity(1, mod_id = "mod_1"))),
-      filter_key = NULL
+      filter_key = NULL,
+      enable_dataset_filter = TRUE
     )
 
     testServer(app_server_test(testing_options), {
@@ -119,7 +124,8 @@ test_that(
     testing_options <- list(
       data = list(),
       module_info = resolve_module_list(list()),
-      filter_key = NULL
+      filter_key = NULL,
+      enable_dataset_filter = TRUE
     )
     shiny::testServer(app_server_test(testing_options), {
     }) %>%
@@ -157,7 +163,8 @@ test_that(
       data = datasets,
       filter_data = "a",
       module_info = resolve_module_list(list()),
-      filter_key = "mpg"
+      filter_key = "mpg",
+      enable_dataset_filter = TRUE
     )
 
     withr::local_locale(.new = list("LC_TIME" = "en_US.UTF-8"))
@@ -195,7 +202,8 @@ test_that(
       data = datasets,
       filter_data = "a",
       module_info = resolve_module_list(list()),
-      filter_key = "mpg"
+      filter_key = "mpg",
+      enable_dataset_filter = TRUE
     )
 
     withr::local_locale(.new = list("LC_TIME" = "en_US.UTF-8"))
@@ -236,7 +244,8 @@ test_that(
       data = datasets,
       filter_data = "a",
       module_info = resolve_module_list(list()),
-      filter_key = "mpg"
+      filter_key = "mpg",
+      enable_dataset_filter = TRUE
     )
 
     withr::local_locale(.new = list("LC_TIME" = "en_US.UTF-8"))
@@ -269,7 +278,8 @@ test_that(
       filter_data = "a",
       module_info = resolve_module_list(list()),
       filter_key = "mpg",
-      reload_period = lubridate::duration(1, "seconds")
+      reload_period = lubridate::duration(1, "seconds"),
+      enable_dataset_filter = TRUE
     )
     withr::with_dir(tempdir(), {
       system2(command = "touch", args = c("restart.txt"), stdout = TRUE)